requirements.txt
flask
werkzeug
fastapi
uvicorn[standard]
Pillow
python-multipart
python-magic
pymupdf